requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
google-cloud-firestore==2.11.1
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
AnueCrawler
google-cloud-logging==3.8.0